spec: opt-in line coverage for .lic files
`COVERAGE=1 bundle exec rspec` now produces a real coverage report in
coverage/ (already gitignored). Three things were in the way:

1. The specs never require the .lic files -- load_lic_class evals a slice
   of them -- and Ruby only measures eval'd source when Coverage is set up
   with eval: true. Without that the report is 0% covered, which is what
   made it look like coverage simply did not work here.

2. The extractors built the path as "<root>/spec/../foo.lic", whose
   project-relative form is "spec/../foo.lic" -- which SimpleCov's default
   test-directory filter discards. Expanding the path via the new lic_path
   helper makes the default filters correct, so no filter surgery is
   needed. It also means backtraces point at the file's real path.

3. Ruby allocates a file's eval-coverage line array on the first eval
   attributed to that filename and never grows it. A multi-class .lic is
   eval'd once per class, so any class starting past the end of the
   first-eval'd one was silently dropped -- not counted as missed, just
   absent. combat-trainer.lic reported 2 of its 11 classes (791 relevant
   lines); it now reports all 11 (3724). prime_lic_coverage sizes the array
   once per file, and is a no-op unless Coverage is running.

While wiring this up: automap_spec and moonwatch_spec each defined their
own top-level load_lic_module, and dependency_spec built its own path.
Two identical top-level defs in different spec files is the load-order
hazard spec_helper's own header warns about, so load_lic_module moves to
spec_helper and the three specs use the shared extractors. That also
stopped automap.lic and dependency.lic being dropped from the report and
moonwatch.lic losing its five modules.

load_lic_constant now passes the constant's real line number to eval
instead of defaulting to line 1.

Suite is unchanged at 2922 examples with and without COVERAGE set.
