+++ Gemfile
@@ -30,6 +30,7 @@ source "https://rubygems.org"
 git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 group :development do
+  gem "simplecov"
   gem "rspec"
   gem "rubocop"
   gem "guard"
